# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for the Omoikane CMS. Public endpoints do not require authentication; protected endpoints accept a Bearer token (or the session cookie set by login). Admin-only endpoints require an admin role.",
        "title": "Omoikane API",
        "contact": {
            "name": "Omoikane"
        },
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/trash": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns soft-deleted pages, users, posts, media, contacts, messages, tags and categories.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trash"
                ],
                "summary": "List trash items",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_handlers.TrashItem"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Hard-deletes all trash items, or only the given entity's items when ?entity= is provided.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trash"
                ],
                "summary": "Empty trash",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Only empty this entity type (page, user, post, media, contact, message, tag, category)",
                        "name": "entity",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    }
                }
            }
        },
        "/trash/count": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the total number of soft-deleted items across all entities.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trash"
                ],
                "summary": "Trash count",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "integer",
                                "format": "int64"
                            }
                        }
                    }
                }
            }
        },
        "/trash/{entity}/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Hard-deletes an item by entity type and ID. Media files are removed from disk.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trash"
                ],
                "summary": "Permanently delete trash item",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Entity type (page, user, post, media, contact, message, tag, category)",
                        "name": "entity",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/trash/{entity}/{id}/restore": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Restores a soft-deleted item by entity type and ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trash"
                ],
                "summary": "Restore trash item",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Entity type (page, user, post, media, contact, message, tag, category)",
                        "name": "entity",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_handlers.TrashItem": {
            "type": "object",
            "properties": {
                "deletedAt": {
                    "type": "string"
                },
                "entity": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}